Update README with examples and re-organize
