+++ examples/flux-chat/js/flux.js
@@ -1,3 +1,4 @@
+// create the Nuclear reactor instance, this will act as our dispatcher and interface for data fetching
 var Nuclear = require('nuclear-js')
 
 module.exports = new Nuclear.Reactor({
