+++ examples/flux-chat/js/modules/chat/action-types.js
