+++ examples/flux-chat/js/modules/chat/actions.js
