+++ examples/flux-chat/js/modules/chat/getters.js
