+++ examples/flux-chat/js/modules/chat/index.js
