+++ examples/flux-chat/js/modules/chat/stores/current-thread-id-store.js
