+++ examples/flux-chat/js/modules/chat/stores/thread-store.js
